15. Composition
This is the one where we stop repeating ourselves again and again and again.

So far.

We've repeated ourselves quite a bit.

The home page and the about page look very, very similar and we sort of duplicated effort.

So I want to introduce an umbraco concept called composition.

And if you're in the programming game, you'll this is another name for sort of inheritance.

So what we can do is we can define a document type and then we can inherit from it in another document

type.

And let's see how that might reduce our duplication, because we all know that repetition and duplication

are sort of evil.

I'm going to log in to the back office.

And go to settings and just highlight what I'm talking about.

So I've got the about page which has content and message and I've got the home page which has got content

and message.

So that seems to me that they're doing sort of similar things.

So why don't I put the message in sort of a reusable object?

So now I'm going to create a document type.

But I'm going to create one without a template.

So really, I'm only concerned with this holding data.

I'm not concerned about a view.

So I'm going to go ahead and create a document type without a template.

And I'm going to call this document type page content.

And this will hold the page content, obviously.

Again, it's that difficult decision about what icon to use.

It's just so difficult.

Shall I use a duck?

I'll use a document.

Yeah, really spend some time on choosing those icons.

And again, I get to add a group, so I was quite happy with it being called content.

And again, I can add a property.

So let me go ahead and call this page content text.

This will be the the text message displayed on the screen.

And I'm going to still use a text box.

Keep the defaults.

Keep the defaults there again, submit.

Great.

So now I've got this document type called Page Content.

It's got this content section with a text box in.

I'm going to pop over to permissions.

And I'm going to choose this special one here called element type.

Now, you would only ever use this element type.

Well, the situations I can think of anyway, if this document type will never appear in the content

tree.

So if I flip back to content.

Sorry.

I'll just save that.

Say that.

Flip back to content.

So this is what I mean.

If you're never going to create something that's going to be in this tree.

Then go ahead and set it as an element type.

So let's go back to page content and permissions.

And we've chose is an element type.

Great.

So I said that composition was sort of another name for inheritance.

So if I'm going back to my home page, how can I make my home page?

Inherit.

Well, Oh, there's this button here called compositions, so if I choose that.

I can inherit from the about page if I want to, but I've just gone to the trouble of creating this

page content, so I'll tick that.

And press submit.

Now, can you see?

Because I already had a section called Content.

It sort of merged it in.

And now I've got two items on my home page.

I've got my original Hello world text and I've got this new item which I can't click on because I've

inherited it.

So I don't think I need Helloworld text anymore, so I can certainly put that in the trash can.

And save that.

Okay.

The trouble is, though, that this text box is called page content text, and I'm pretty sure if I

go to my template.

It's still looking for Hello world text.

So if I just update this to be page content, text and save.

And now go back to the content tab.

And the home page.

We'll see that my previous text box is gone and I've got this new text box and I'll now say Hello world

from the composition.

Okay, let me save that.

And I'll use my cheeky little shortcut and have a look at what the page is like.

Hello world from the composition.

So as a challenge, why not you try updating the about page to use the same composition?