Detener 16. Composition - adding to the About page
So I hope you accepted the challenge about adding the composition to the about page.

Let me just go through it.

I'm going to log in again to the back office.

I'm going to go to the settings.

Open up my document types.

Go to about.

Compositions.

Yes.

Add in the page content.

Let's get rid of the existing text box there.

Save that.

Need to go and update the about template.

Instead of saying message text, it needs to say page, content, text.

And remember, you could also have got that value from our chooser field.

Okay, save that.

Now, since this was a new text box, I will need to go into my content section to the about page.

And add the content.

Composition style, of course, with an exclamation.

Save and publish that.

Use my handy link in the info tab to open up the about page.

And there we go.

So I've introduced you to compositions in Umbraco, which to me seems a little bit like classic inheritance

in object oriented programming.

I should really introduce another concept at this stage.

While we're dealing with duplication and repetition.

And in the next section we will talk about master Pages.